Add support for multiple download links for multiple OSes for software

+++ generatePageData/index.js
@@ -1,8 +1,6 @@
 const fs = require('fs')
 
-if (!fs.existsSync('./docs/.vuepress/public/pageData')) fs.mkdirSync('./docs/.vuepress/public/pageData', { recursive: true })
-
-fs.rmSync('./docs/.vuepress/public/pageData', { recursive: true });
+if (fs.existsSync('./docs/.vuepress/public/pageData')) fs.rmSync('./docs/.vuepress/public/pageData', { recursive: true });
 fs.mkdirSync('./docs/.vuepress/public/pageData', { recursive: true })
 
 require('./generateFirmware')
